adding react routing and understanding how it works
